requirements.txt
Django==1.11.10
mysql-connector-python==8.0.6


